+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+### [ita-challenges-frontend-3.28.0-RELEASE] - 2026-02-19
+### Added
+- Bookmark button UI in challenge card component
 
 ### [ita-challenges-frontend-3.27.4-RELEASE] - 2026-02-18
 
@@ -23,6 +26,7 @@ and this project adheres to
 
 ### Fixed
 - Fixed environment.ts and environment.prod.ts route for solution submission and retrieval.
+
 
 ### [ita-challenges-frontend-3.27.0-RELEASE] - 2026-02-16